fix: move entrypoint checker to config; cyclic imports
