+++ src/worker/batchSubmitter/index.ts
@@ -4,8 +4,7 @@ import { BatchSubmitter } from './batchSubmitter'
 import { initServer, finalizeServer, initMetricsServer } from '../../loader'
 import { batchController, metricsController } from '../../controller'
 import { once } from 'lodash'
-import { config } from '../../config'
-import {isInvokedFromEntrypoint} from "../../entrypoint"
+import { config, isInvokedFromEntrypoint } from '../../config'
 
 let jobs: BatchSubmitter[] = []
